Prepare for embedding libtcc1.a to r3
A new cmake option TCC_EMBEDDED_IN_R3 is introduced, when it's set to
true, tcc_relocate will stop looking for libtcc1.a.

All of the functions that are exported by libtcc1.a are stored in a symbol
table libtcc1_symbols[], which can be added to the compiled code by
calling tcc_add_symbols. All of the functions are defined as "static" to
avoid name conflicts.